adding modifications of 0.4.2-30

+++ src/ags/audio/ags_recycling.h
@@ -75,11 +75,11 @@ void ags_recycling_remove_audio_signal(AgsRecycling *recycling,
 
 void ags_recycling_create_audio_signal_with_defaults(AgsRecycling *recycling,
 						     AgsAudioSignal *audio_signal,
-						     guint delay, guint attack);
+						     gdouble delay, guint attack);
 void ags_recycling_create_audio_signal_with_frame_count(AgsRecycling *recycling,
 							AgsAudioSignal *audio_signal,
 							guint frame_count,
-							guint delay, guint attack);
+							gdouble delay, guint attack);
 
 gint ags_recycling_position(AgsRecycling *start_recycling, AgsRecycling *end_region,
 			    AgsRecycling *recycling);
