+++ src/ags/audio/recall/ags_delay_audio_run.h
@@ -52,35 +52,35 @@ struct _AgsDelayAudioRunClass
   AgsRecallAudioRunClass recall_audio_run;
 
   void (*notation_alloc_output)(AgsDelayAudioRun *delay_audio_run, guint run_order,
-				guint delay, guint attack);
+				gdouble delay, guint attack);
   void (*notation_alloc_input)(AgsDelayAudioRun *delay_audio_run, guint run_order,
-			       guint delay, guint attack);
+			       gdouble delay, guint attack);
   void (*notation_count)(AgsDelayAudioRun *delay_audio_run, guint run_order,
-			 guint delay, guint attack);
+			 gdouble delay, guint attack);
 
   void (*sequencer_alloc_output)(AgsDelayAudioRun *delay_audio_run, guint run_order,
-				 guint delay, guint attack);
+				 gdouble delay, guint attack);
   void (*sequencer_alloc_input)(AgsDelayAudioRun *delay_audio_run, guint run_order,
-				guint delay, guint attack);
+				gdouble delay, guint attack);
   void (*sequencer_count)(AgsDelayAudioRun *delay_audio_run, guint run_order,
-			  guint delay, guint attack);
+			  gdouble delay, guint attack);
 };
 
 GType ags_delay_audio_run_get_type();
 
 void ags_delay_audio_run_notation_alloc_output(AgsDelayAudioRun *delay_audio_run, guint run_order,
-					       guint delay, guint attack);
+					       gdouble delay, guint attack);
 void ags_delay_audio_run_notation_alloc_input(AgsDelayAudioRun *delay_audio_run, guint run_order,
-					      guint delay, guint attack);
+					      gdouble delay, guint attack);
 void ags_delay_audio_run_notation_count(AgsDelayAudioRun *delay_audio_run, guint run_order,
-					guint delay, guint attack);
+					gdouble delay, guint attack);
 
 void ags_delay_audio_run_sequencer_alloc_output(AgsDelayAudioRun *delay_audio_run, guint run_order,
-						guint delay, guint attack);
+						gdouble delay, guint attack);
 void ags_delay_audio_run_sequencer_alloc_input(AgsDelayAudioRun *delay_audio_run, guint run_order,
-					       guint delay, guint attack);
+					       gdouble delay, guint attack);
 void ags_delay_audio_run_sequencer_count(AgsDelayAudioRun *delay_audio_run, guint run_order,
-					 guint delay, guint attack);
+					 gdouble delay, guint attack);
 
 AgsDelayAudioRun* ags_delay_audio_run_new();
